+++ src/s1/tcf-2.0-loader.js
@@ -16,6 +16,12 @@
 			polyfillSrc = 'polyfillSrc',
 			shouldPolyfill = !window.Promise || !window.fetch || !window.Symbol;
 
+		// Test for specific functionality for some polyfills
+		if (!shouldPolyfill) {
+			var promiseTest = new Promise(function () {});
+			shouldPolyfill = !promiseTest.finally;
+		}
+
 		function loadScript(src, done) {
 			var js = document.createElement('script');
 			js.src = src;